+++ packages/builder-util/CHANGELOG.md
@@ -1,5 +1,21 @@
 # builder-util
 
+## 27.0.0-alpha.8
+
+### Patch Changes
+
+<details><summary>Updated 1 dependency</summary>
+
+<small>
+
+[`cf39086`](https://github.com/electron-userland/electron-builder/commit/cf39086fbb71e34d1fef0359a026697153e6ee3b)
+
+</small>
+
+- `builder-util-runtime@10.0.0-alpha.7`
+
+</details>
+
 ## 27.0.0-alpha.7
 
 ### Patch Changes